updated the db.ts renamed  from index to db. removed unnessary imports
