+++ server/src/database/seeders/sync.ts
@@ -3,7 +3,7 @@ export { };
 
 import { Sequelize } from "sequelize";
 import dotenv from 'dotenv';
-import db from "../models";
+import db from "../models/db";
 
 dotenv.config();
